fixed typo: identithub -> identityhub
