+++ extensions/did/local-did-publisher/src/main/java/org/eclipse/edc/identityhub/publisher/did/local/DidDocumentObservableImpl.java
@@ -14,8 +14,8 @@
 
 package org.eclipse.edc.identityhub.publisher.did.local;
 
-import org.eclipse.edc.identithub.spi.did.events.DidDocumentListener;
-import org.eclipse.edc.identithub.spi.did.events.DidDocumentObservable;
+import org.eclipse.edc.identityhub.spi.did.events.DidDocumentListener;
+import org.eclipse.edc.identityhub.spi.did.events.DidDocumentObservable;
 import org.eclipse.edc.spi.observe.ObservableImpl;
 
 public class DidDocumentObservableImpl extends ObservableImpl<DidDocumentListener> implements DidDocumentObservable {
