+++ spi/did-spi/src/main/java/org/eclipse/edc/identityhub/spi/did/package-info.java
@@ -12,12 +12,12 @@
  *
  */
 
-package org.eclipse.edc.identithub.spi.did.events;
-
-import org.eclipse.edc.spi.observe.Observable;
 
 /**
- * Manages and invokes {@link DidDocumentListener}s when a state change related to a DID document resource has happened.
+ * This module declares several service interfaces and extension points that are required to host, publish and store
+ * DID documents.
  */
-public interface DidDocumentObservable extends Observable<DidDocumentListener> {
-}
+@Spi(value = "Identity Hub DID services")
+package org.eclipse.edc.identityhub.spi.did;
+
+import org.eclipse.edc.runtime.metamodel.annotation.Spi;
