+++ spi/did-spi/src/test/java/org/eclipse/edc/identityhub/spi/did/DidWebParserTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.identithub.spi.did;
+package org.eclipse.edc.identityhub.spi.did;
 
 import org.junit.jupiter.api.Test;
